1538 - Add the possibility to update happening participation without removing or updating question

+++ src/Application/Command/Happening/ParticipateHandler.php
@@ -178,7 +178,7 @@ class ParticipateHandler
             }
         }
 
-        if (true === $participate->happening->isQuestionAllowed()) {
+        if (true === $participate->happening->isQuestionAllowed() && $participate->updateQuestion) {
             // Remove previous question
             $this->questionRepository->removeQuestionFromUserForHappening(
                 $participate->createdBy,
